Tomorrowland
so far decentraland is the better version of the metaverse. compared to others the graphics quality is beyond amazing which attracted fashion shows, events and a lot more entertainments along.
A small version of  Tomorrowland, a land of piece love and music is the only thing that is missing the metaverse.